Added tearDown method in ConstructorControllerTest and added a helper method getSource in ControllerTestUtils
